services/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Gunnar Kistner <gunnar.kistner@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.96.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
strawberry-graphql = {extras = ["fastapi"], version = "^0.181.0"}
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
